Assume profit is a result of buy + sell only
ccgains assumption is that if you deposited some quote currency and sold
it for base currency, all obtained amount will be considered as a
profit. Because we primarily interesting in evaluating strategy
performance, we considering profit as a result of buy+sell only. So such
deposited and then sold currency will be counted for 0 profit.